Added Highscores

+++ db/migrate/20220723073125_create_scores.rb
@@ -1,7 +1,7 @@
 class CreateScores < ActiveRecord::Migration[5.2]
   def change
     create_table :scores do |t|
-      t.integer :point
+      t.float :point
 
       t.timestamps
     end
